backend/requirements.txt
# Web framework
fastapi>=0.110.0
uvicorn[standard]>=0.29.0

# Database
asyncpg>=0.29.0
psycopg[binary]>=3.1.0

# Migrations
alembic>=1.13.0

# Validation & Settings
pydantic>=2.6.0
pydantic-settings>=2.2.0

# Auth
python-jose[cryptography]>=3.3.0
bcrypt==4.0.1

# Env
python-dotenv>=1.0.0

# File uploads (нужен для UploadFile в FastAPI)
python-multipart>=0.0.9

# Async HTTP (геокодинг адресов для карты)
httpx>=0.27.0

# Async file operations
aiofiles>=23.2.0